Extract the term-order tiebreaker into a reusable trait
Move executeTermQueryWithStableOrder() out of AbstractTaxonomyTypeAPI into
a StableTaxonomyTermOrderTrait, so the same deterministic term-ID
tiebreaker can be reused by other term-fetching APIs without duplicating
the terms_clauses logic.
